# Cache disk ID for a short while