## Motivation and Context

`*posix.getDiskID()` takes up to 30% of all CPU due to the `os.Stat` call on `GET` calls. (1MB objects)

Only check time on the format JSON at most every second.

Also `fileInfo` was accessed without proper lock.

Before:
```
Operation: GET - Concurrency: 12
* Average: 1333.97 MB/s, 1365.99 obj/s, 1365.98 ops ended/s (4m59.975s)
* First Byte: Average: 7.801487ms, Median: 7.9974ms, Best: 1.9822ms, Worst: 110.0021ms

Aggregated, split into 299 x 1s time segments:
* Fastest: 1453.50 MB/s, 1488.38 obj/s, 1492.00 ops ended/s (1s)
* 50% Median: 1360.47 MB/s, 1393.12 obj/s, 1393.00 ops ended/s (1s)
* Slowest: 978.68 MB/s, 1002.17 obj/s, 1004.00 ops ended/s (1s)
```

After:
```
Operation: GET - Concurrency: 12
* Average: 1759.10 MB/s, 1759.10 obj/s, 1759.09 ops ended/s (4m59.949s)
* First Byte: Average: 5.953116ms, Median: 5.9967ms, Best: 992.2µs, Worst: 166.9996ms

Aggregated, split into 299 x 1s time segments:
* Fastest: 1935.89 MB/s, 1935.89 obj/s, 1936.00 ops ended/s (1s)
* 50% Median: 1795.35 MB/s, 1795.35 obj/s, 1798.00 ops ended/s (1s)
* Slowest: 914.17 MB/s, 914.17 obj/s, 915.00 ops ended/s (1s)
```

![image](https://user-images.githubusercontent.com/5663952/69861127-fbacf800-1297-11ea-872b-4767e0319d1c.png)

## How to test this PR?

Needs to be reviewed by someone who knows if my assumptions above holds.

Alternatively we could add a check only every 10 seconds.

## Types of changes
- [x] Optimization