@klauspost 2019-11-22
CI failure:
```
[91m# cd .; git clone -- https://go.googlesource.com/lint /go/src/golang.org/x/lint
3769
Cloning into '/go/src/golang.org/x/lint'...
3770
fatal: unable to access 'https://go.googlesource.com/lint/': Failed to connect to go.googlesource.com port 443: Connection timed out
```