@harshavardhana 2019-11-22
@klauspost there is a bug here you have removed SameFile() which is needed for us to check the inode has changed i.e some other disk is not mounted automatically underneath, this happens quite commonly in situations when they are hot swapping etc.  We need to make sure that didn't happen , so removing that code adds room for situations such as these. 

Btw it's a Linux only check, since Windows never allows disks to be swapped in that manner.