@klauspost 2019-11-23
@harshavardhana Makes sense. I have updated the PR so we check every second. PTAL.

Windows can hotswap too. Scary, I know! :)